Stockfish 16

+++ Chess/eval.h
@@ -1,7 +1,7 @@
 #ifndef EVAL_H
 #define EVAL_H
 
-using value = int;
+using value = long long;
 using bitboard = std::bitset<64>;
 
 #endif // !EVAL_H
